+++ Spot.IM-Development/SpotImCore/Core/Models/Configs/Configuration/SPConfigurationSDKStatus.swift
@@ -9,7 +9,7 @@
 import Foundation
 
 struct SPConfigurationSDKStatus: Decodable {
-    let enabled: Bool? 
+    let enabled: Bool?
     let locale: String?
     let realtimeEnabled: Bool?
     let loginUiEnabled: Bool?