+++ Spot.IM-Development/CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Only call ab_test/ads/user if SDK is enabled in the main config
+- Call user/data endpoint only when needed
 
 ## [1.0.7] - 2020-08-20
 ### Changed